Moves a functions into a new file + Fixes the unnamed numbers
