+++ functions/venue.js
@@ -1,7 +1,6 @@
 const admin = require("firebase-admin");
 const functions = require("firebase-functions");
 const { HttpsError } = require("firebase-functions/lib/providers/https");
-const { chunk, sampleSize } = require("lodash");
 
 const { addAdmin, removeAdmin } = require("./src/api/roles");
 
@@ -930,53 +929,3 @@ exports.setVenueLiveStatus = functions.https.onCall(async (data, context) => {
 
   await admin.firestore().collection("venues").doc(data.venueId).update(update);
 });
-
-exports.aggregateUsersLocationsInVenue = functions.pubsub
-  .schedule("every 5 minutes")
-  .onRun(async () => {
-    const venuesPromise = admin
-      .firestore()
-      .collection("venues")
-      .get()
-      .then((snapshot) =>
-        snapshot.docs.map((doc) => ({ ...doc.data(), id: doc.id }))
-      );
-
-    const usersPromise = admin
-      .firestore()
-      .collection("users")
-      .get()
-      .then((snapshot) =>
-        snapshot.docs.map((doc) => ({ ...doc.data(), id: doc.id }))
-      );
-
-    const [users, venues] = await Promise.all([usersPromise, venuesPromise]);
-
-    // NOTE: Chunk users into batch commit to faster the operation
-    chunk(venues, 250).forEach((venuesChunk) => {
-      const batch = admin.firestore().batch();
-
-      venuesChunk.forEach((venue) => {
-        const recentVenueUsers = users.filter(
-          (user) =>
-            user.lastVenueIdSeenIn && user.lastVenueIdSeenIn.includes(venue.id)
-        );
-
-        const recentVenueUsersCount = recentVenueUsers.length;
-
-        const venueRef = admin.firestore().collection("venues").doc(venue.id);
-
-        batch.update(venueRef, {
-          recentUserCount: recentVenueUsersCount,
-          recentUsersSample: sampleSize(
-            recentVenueUsers,
-            venue.recentUsersSampleSize ?? 6
-          ),
-        });
-      });
-
-      batch.commit();
-    });
-
-    return null;
-  });
